front: create new flower entry form and required service back: create FlowerController and EntryService to validate and create new entry

+++ natura-server/src/main/java/com/natura/web/server/controllers/UserController.java
@@ -45,7 +45,7 @@ public class UserController {
     }
 
     @GetMapping(path="/{id}")
-    public @ResponseBody User getUser(@PathVariable Integer id) {
+    public @ResponseBody User getUser(@PathVariable Long id) {
 
         return userRepository.findById(id).orElse(null);
     }
